+++ dipy/tracking/local_tracking.py
