+++ dipy/workflows/tracking.py
@@ -12,8 +12,8 @@ from dipy.io.peaks import load_peaks
 from dipy.io.stateful_tractogram import Space, StatefulTractogram
 from dipy.io.streamline import save_tractogram
 from dipy.tracking import utils
-from dipy.tracking.localtracking import (LocalTracking,
-                                         ParticleFilteringTracking)
+from dipy.tracking.local_tracking import (LocalTracking,
+                                          ParticleFilteringTracking)
 from dipy.tracking.tissue_classifier import (BinaryTissueClassifier,
                                              CmcTissueClassifier,
                                              ThresholdTissueClassifier)
