+++ doc/examples/tracking_probabilistic.py
@@ -26,7 +26,7 @@ from dipy.data import read_stanford_labels
 from dipy.reconst.csdeconv import (ConstrainedSphericalDeconvModel,
                                    auto_response)
 from dipy.tracking import utils
-from dipy.tracking.localtracking import LocalTracking
+from dipy.tracking.local_tracking import LocalTracking
 from dipy.tracking.streamline import Streamlines
 from dipy.tracking.tissue_classifier import ThresholdTissueClassifier
 from dipy.viz import window, actor, colormap, has_fury
